# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nodelo Backend API",
    "description": "RESTful API for Nodelo website backend. Includes contact form submissions, authentication, real-time chat with file uploads, and admin management endpoints. Features Socket.io for real-time messaging and Cloudinary for file storage.",
    "version": "1.0.0",
    "contact": {
      "name": "Nodelo Team",
      "email": "hello@nodelo.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5050",
      "description": "Development server"
    },
    {
      "url": "https://api.nodelo.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/chats/from-contact": {
      "post": {
        "tags": ["Chats"],
        "summary": "Create chat from contact",
        "description": "Create a new chat conversation from an existing contact submission. Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["contactId"],
                "properties": {
                  "contactId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "adminId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439012",
                    "nullable": true,
                    "description": "Admin ID to assign. Defaults to authenticated admin."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chat created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Chat created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Chat"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contact or admin not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chats": {
      "post": {
        "tags": ["Chats"],
        "summary": "Create new chat",
        "description": "Create a new chat conversation. Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userName", "userEmail"],
                "properties": {
                  "userName": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "userEmail": {
                    "type": "string",
                    "format": "email",
                    "example": "john@example.com"
                  },
                  "adminId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439012",
                    "nullable": true,
                    "description": "Admin ID to assign. Defaults to authenticated admin."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chat created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Chat"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Chats"],
        "summary": "Get all chats",
        "description": "Retrieve all chat conversations with pagination and filtering. Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by chat status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["active", "resolved", "archived"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of chats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Chat"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{id}": {
      "get": {
        "tags": ["Chats"],
        "summary": "Get chat with messages",
        "description": "Retrieve a single chat conversation with all messages. Supports optional authentication for users.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for messages",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Messages per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat details with messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "chat": {
                          "$ref": "#/components/schemas/Chat"
                        },
                        "messages": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Message"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/message": {
      "post": {
        "tags": ["Chats"],
        "summary": "Send message (public)",
        "description": "Send a message to a chat. Supports file uploads (images, videos, PDFs). This is a public endpoint for users.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["chatId", "senderName", "userEmail"],
                "properties": {
                  "chatId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "content": {
                    "type": "string",
                    "example": "Hello, I need help with my project"
                  },
                  "senderName": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "userEmail": {
                    "type": "string",
                    "format": "email",
                    "example": "john@example.com"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Files to upload (max 5, 50MB each). Supported: images, videos, PDFs"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["chatId"],
                "properties": {
                  "chatId": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "senderName": {
                    "type": "string"
                  },
                  "userEmail": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{id}/message": {
      "post": {
        "tags": ["Chats"],
        "summary": "Send message to specific chat",
        "description": "Send a message to a specific chat. Supports file uploads. Authentication optional (for users).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent successfully"
          }
        }
      }
    },
    "/api/chats/reply-contact": {
      "post": {
        "tags": ["Chats"],
        "summary": "Admin reply to contact",
        "description": "Reply to a contact form submission. Creates chat if needed and sends email notification to user. Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["contactId", "message"],
                "properties": {
                  "contactId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "message": {
                    "type": "string",
                    "example": "Thank you for your inquiry. We'll get back to you soon."
                  },
                  "chatId": {
                    "type": "string",
                    "nullable": true,
                    "description": "Existing chat ID if chat already exists"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reply sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Reply sent successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "$ref": "#/components/schemas/Message"
                        },
                        "chat": {
                          "$ref": "#/components/schemas/Chat"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/user/{userEmail}": {
      "get": {
        "tags": ["Chats"],
        "summary": "Get user's chats",
        "description": "Get all chats for a specific user by email address.",
        "parameters": [
          {
            "name": "userEmail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User's chats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Chat"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/mark-read": {
      "post": {
        "tags": ["Chats"],
        "summary": "Mark messages as read",
        "description": "Mark all unread messages in a chat as read. Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["chatId"],
                "properties": {
                  "chatId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Messages marked as read"
          }
        }
      }
    },
    "/api/chats/{id}/status": {
      "patch": {
        "tags": ["Chats"],
        "summary": "Update chat status",
        "description": "Update the status of a chat (active, resolved, archived). Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["active", "resolved", "archived"],
                    "example": "resolved"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Chat"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login endpoint. Include in Authorization header as: Bearer <token>"
      }
    },
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "pages": {
            "type": "integer",
            "example": 5
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message description"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "example": ["Email is required", "Password must be at least 6 characters"]
          },
          "message": {
            "type": "string",
            "nullable": true,
            "example": "Additional error details (development only)"
          }
        }
      },
      "Chat": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "contactId": {
            "type": "string",
            "nullable": true,
            "example": "507f1f77bcf86cd799439012",
            "description": "Linked contact submission ID"
          },
          "userId": {
            "type": "string",
            "example": "user@example.com",
            "description": "User identifier (usually email)"
          },
          "userName": {
            "type": "string",
            "example": "John Doe"
          },
          "userEmail": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "adminId": {
            "type": "string",
            "nullable": true,
            "example": "507f1f77bcf86cd799439013",
            "description": "Assigned admin user ID"
          },
          "adminName": {
            "type": "string",
            "nullable": true,
            "example": "Admin User",
            "description": "Name of the admin handling this chat"
          },
          "status": {
            "type": "string",
            "enum": ["active", "resolved", "archived"],
            "default": "active",
            "example": "active"
          },
          "lastMessageAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "lastMessageBy": {
            "type": "string",
            "enum": ["user", "admin"],
            "example": "user"
          },
          "unreadCount": {
            "type": "object",
            "properties": {
              "user": {
                "type": "integer",
                "example": 2
              },
              "admin": {
                "type": "integer",
                "example": 5
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "chatId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439012"
          },
          "senderType": {
            "type": "string",
            "enum": ["user", "admin"],
            "example": "user"
          },
          "senderId": {
            "type": "string",
            "nullable": true,
            "example": "507f1f77bcf86cd799439013",
            "description": "Admin user ID if senderType is admin"
          },
          "senderName": {
            "type": "string",
            "example": "John Doe"
          },
          "content": {
            "type": "string",
            "example": "Hello, I need help with my project",
            "nullable": true
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "format": "uri",
                  "example": "https://res.cloudinary.com/.../image.jpg"
                },
                "type": {
                  "type": "string",
                  "enum": ["image", "video", "pdf", "document"],
                  "example": "image"
                },
                "fileName": {
                  "type": "string",
                  "example": "project-screenshot.png"
                },
                "fileSize": {
                  "type": "integer",
                  "example": 1024000,
                  "description": "File size in bytes"
                },
                "publicId": {
                  "type": "string",
                  "example": "nodelo/messages/abc123",
                  "description": "Cloudinary public ID"
                }
              }
            },
            "description": "File attachments (images, videos, PDFs)"
          },
          "isRead": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2024-01-15T10:35:00.000Z"
          },
          "isReplyToContact": {
            "type": "boolean",
            "default": false,
            "example": false,
            "description": "True if this is an admin reply to a contact form"
          },
          "emailSent": {
            "type": "boolean",
            "default": false,
            "example": true,
            "description": "True if email notification was sent"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00.000Z"
          }
        }
      }
    }
  }
}